+++ plugins/proclist/Proclist.cpp
@@ -175,7 +175,7 @@ static void dynamic_bind(void)
 
 int Message(unsigned Flags, wchar_t *HelpTopic, LPCTSTR*Items, int nItems, int nButtons)
 {
-	return Info.Message(&MainGuid, nullptr, Flags, HelpTopic, Items, nItems, nButtons);
+	return (int)Info.Message(&MainGuid, nullptr, Flags, HelpTopic, Items, nItems, nButtons);
 }
 
 //-----------------------------------------------------------------------------
